+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.7.8...HEAD) - TBD
 
+### :bug: Fixed
+
+- Unhandled `PermissionError` when loading config file.
+
 ## [4.7.8](https://github.com/schemathesis/schemathesis/compare/v4.7.7...v4.7.8) - 2025-12-28
 
 ### :wrench: Changed